# steven/helix-prototypes/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalCommand": {
        "additionalProperties": false,
        "properties": {
          "meaning": {
            "maxLength": 200,
            "minLength": 4,
            "title": "Meaning",
            "type": "string"
          },
          "reviewer": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Reviewer",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ApprovalRole"
          }
        },
        "required": [
          "role",
          "reviewer",
          "meaning"
        ],
        "title": "ApprovalCommand",
        "type": "object"
      },
      "ApprovalRole": {
        "enum": [
          "pathologist",
          "peer_reviewer",
          "qau",
          "study_director"
        ],
        "title": "ApprovalRole",
        "type": "string"
      },
...
      "DataValidationCommand": {
        "additionalProperties": false,
        "properties": {
          "actor": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Actor",
            "type": "string"
          },
          "idempotency_key": {
            "maxLength": 160,
            "minLength": 8,
            "title": "Idempotency Key",
            "type": "string"
          },
          "package_id": {
            "default": "validation.body_weight",
            "maxLength": 120,
            "minLength": 1,
            "title": "Package Id",
            "type": "string"
          }
        },
        "required": [
          "actor",
          "idempotency_key"
        ],
        "title": "DataValidationCommand",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "A synthetic pattern-testing API. It does not certify GLP, Part 11, SEND, eCTD, or FDA acceptance.",
    "title": "HELIX synthetic nonclinical workflow API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/studies/{study_id}/approvals": {
      "post": {
        "operationId": "record_approval_api_v1_studies__study_id__approvals_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovalCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Approval",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/studies/{study_id}/claims/{claim_id}/evidence": {
      "get": {
        "operationId": "get_evidence_api_v1_studies__study_id__claims__claim_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceChain"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "evidence"
        ]
      }
    },
    "/api/v1/studies/{study_id}/data-validation-packages": {
      "post": {
        "operationId": "run_data_validation_api_v1_studies__study_id__data_validation_packages_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataValidationCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataValidationExecution"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Data Validation",
        "tags": [
          "data-validation"
        ]
      }
    },
    "/api/v1/studies/{study_id}/exports": {
      "post": {
        "operationId": "export_package_api_v1_studies__study_id__exports_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Package",
        "tags": [
          "export"
        ]
      }
    },
    "/api/v1/studies/{study_id}/exports/{artifact_id}": {
      "get": {
        "operationId": "download_artifact_api_v1_studies__study_id__exports__artifact_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "artifact_id",
            "required": true,
            "schema": {
              "title": "Artifact Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Artifact",
        "tags": [
          "export"
        ]
      }
    },
    "/api/v1/studies/{study_id}/final-study-approvals": {
      "post": {
        "operationId": "record_final_study_approval_api_v1_studies__study_id__final_study_approvals_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalStudyApprovalCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Record Final Study Approval",
        "tags": [
          "review"
        ]
      }
    },
    "/api/v1/studies/{study_id}/pinned-runs": {
      "post": {
        "operationId": "freeze_run_api_v1_studies__study_id__pinned_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FreezeRunCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PinnedRun"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Freeze Run",
        "tags": [
          "run-plans"
        ]
      }
    },
    "/api/v1/studies/{study_id}/pinned-runs/{run_id}/events": {
      "get": {
        "description": "Live stream of persisted, projected run events for one Pinned Run. Each SSE frame carries `id` (the stable event_id), `event` (the event type), and `data` (a RunEvent JSON object). `Last-Event-ID` (header, or the `last_event_id` query parameter) replays only events after that cursor. If missed events are outside the retained window the server returns HTTP 409 `event_cursor_expired` before opening the stream; refresh GET /workspace and reconnect from `journey.run.latest_event_id`. This stream is not the audit history: `WorkspaceResponse.events` remains the append-only audit record.",
        "operationId": "stream_run_events_api_v1_studies__study_id__pinned_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "last_event_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last Event Id"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "description": "SSE frames whose data is a RunEvent",
                  "type": "string"
                }
              }
            },
            "description": "text/event-stream of RunEvent frames"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvalidEventCursor"
                }
              }
            },
            "description": "Malformed or foreign Last-Event-ID"
          },
          "404": {
            "description": "Unknown study or Pinned Run"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventCursorExpired"
                }
              }
            },
            "description": "Last-Event-ID is outside retention"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream projected run events (Server-Sent Events)",
        "tags": [
          "run-plans"
        ]
      }
    },
    "/api/v1/studies/{study_id}/section-revisions": {
      "post": {
        "operationId": "revise_section_api_v1_studies__study_id__section_revisions_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HumanDirectedRevisionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HumanDirectedRevisionReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revise Section",
        "tags": [
          "section-revisions"
        ]
      }
    },
    "/api/v1/studies/{study_id}/section-runs": {
      "post": {
        "operationId": "run_section_api_v1_studies__study_id__section_runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "title": "Study Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SectionRunCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectionRunReceipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Section",
        "tags": [
          "section-runs"
        ]
      }
    },
...
  }
}